don't extract XML-specific stuff into its own service, just filesystem-specific stuff
