+++ dashboard/test/models/levels/blockly_test.rb
@@ -200,9 +200,9 @@ class BlocklyTest < ActiveSupport::TestCase
 
   test 'converts from and to XML level format' do
     name = 'Test level convert'
-    level = Services::LevelXml.load_custom_level_xml(File.read(File.join(self.class.fixture_path, 'test_level.xml')), Level.new(name: name))
+    level = Services::LevelFiles.load_custom_level_xml(File.read(File.join(self.class.fixture_path, 'test_level.xml')), Level.new(name: name))
     xml = level.to_xml
-    xml2 = Services::LevelXml.load_custom_level_xml(xml, Level.new(name: name.next)).to_xml
+    xml2 = Services::LevelFiles.load_custom_level_xml(xml, Level.new(name: name.next)).to_xml
     level.destroy
     assert_equal xml, xml2
   end
